Move failed reshare tests to own file
